+++ CHANGELOG.md
@@ -335,6 +335,7 @@
 * [X] breeze-serializer 让breeze-serializer也委托breeze-mapper实现并作为默认实现，同时完善相关代码。
 * [X] 更改目录结构。
 * [X] breeze-dsl-command-line 优化代码。
+* [X] breeze-dsl-critic-markup & breeze-dsl-command-line 优化代码。
 * [ ] breeze-core 让集合的`deepQuery`方法当特定的占位符不匹配类型时，回调为普通字符串。 
 * [ ] breeze-unstable 新的集合类型：`Tree`。
 * [ ] breeze-unstable 新的集合类型：`Table`。